I have added CourseRepository,CatalogRepository,CourseController,CourseService,Course classes

+++ src/main/java/com/improved_enrollment_system/improved_enrollment_system/service/AdministratorService.java
@@ -28,7 +28,11 @@ public class AdministratorService {
         return administratorRepository.save(administrator);
     }
 
-    public void deleteAdministrator(Long id) {
-        administratorRepository.deleteById(id);
+    public boolean deleteAdministrator(Long id) {
+        if (administratorRepository.existsById(id)) {
+            administratorRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
